fix: respect superscalar history
Use previous predictions made this cycle in the global history.
This will not be observable until used in Toooba since the ChampSim setup is not superscalar.
Rename GHist to GlobalHistory.
Rename NumGHistBits to NumGlobalHistoryBits.
Use mkRegU instead of mkReg(?). This has no effect.
Update global history in a rule rather than the update method.
Add TODOs relating to mispredictions and incorrect global history.