refactor: make GSelect look parameterised
so that it is ready to become actually parameterisable.
Define Result type as Bool, rather than using Bool and Bits.
Use a vector of results as global history. It would be nice to have more choice over this soon.
Use a (Boolean) value with a hysteresis bit instead of a 2-bit saturation counter